Transaction section completed

+++ Wallet-Service/Wallet-Service/src/main/java/com/example/Walletservice/WalletService/Repository/WalletRepo.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
 public interface WalletRepo extends JpaRepository<Wallet,Integer> {
     Wallet findByUserId(Integer userId);
     @Query("update Wallet w set w.balance = :balance where w.userId = :userId")
-   void updateBalance(Integer userId,Integer balance);
+   void updateBalance(Integer userId,Double balance);
 }
